+++ src/components/DebotsList/index.jsx
@@ -1,8 +1,9 @@
 import React, { Fragment } from 'react';
 import { useSelector } from 'react-redux';
-import { Loader, ControlWithPopup, AddDebot, Environment } from '/src/components';
+import { Loader } from '/src/components';
 import { isCustomScrollBar } from '/src/helpers';
 import DebotListItem from './ListItem';
+import DebotListControls from './ListControls';
 import './index.scss';
 
 const DebotsList = () => {
@@ -31,14 +32,7 @@ const DebotsList = () => {
 
 	return (
 		<div className='debots-list'>
-			<div className='debots-list__controls'>
-				<ControlWithPopup height={310} width={500} name='Add DeBot'>
-					<AddDebot />
-				</ControlWithPopup>
-				<ControlWithPopup height={472} width={660} name='Show Environment'>
-					<Environment />
-				</ControlWithPopup>
-			</div>
+			<DebotListControls />
 			<div className={debotsListClassName}>
 				{!!filteredLocalDebots.length
 					&& <Fragment>
