+++ src/components/Loader/index.scss
@@ -7,7 +7,7 @@
 	max-height: 300px;
 	margin-top: 5vh;
 
-	&__cat {
+	&__content {
 		width: 100%;
 		height: 100%;
 	}
