+++ src/store/actions/environment.js
@@ -1,8 +1,6 @@
 export const SET_ENVIRONMENT_VARIABLE = 'debot/SET_ENVIRONMENT_VARIABLES';
 
-export const setEnvironmentVariables = variables => dispatch => {
-	dispatch({
-		type: SET_ENVIRONMENT_VARIABLE,
-		payload: variables,
-	})
-}
+export const setEnvironmentVariables = variables => ({
+	type: SET_ENVIRONMENT_VARIABLE,
+	payload: variables,
+})
