+++ src/tonClient.js
@@ -2,7 +2,6 @@ import { TonClient } from '@tonclient/core';
 import { libWeb } from '@tonclient/lib-web';
 import { TON_NETWORK_LS_FIELD, MAIN_NETWORK, DEV_NETWORK, FLD_NETWORK } from '/src/constants';
 
-// eslint-disable-next-line react-hooks/rules-of-hooks
 TonClient.useBinaryLibrary(libWeb);
 
 class TonClientController {
